ingestion/pyproject.toml
[project]
name = "noaa-ingestion"
version = "0.1.0"
description = "NOAA GHCN-D ingestion: bulk backfill + Kafka streaming producer"
requires-python = ">=3.11"
dependencies = [
    "clickhouse-connect>=0.7.0",
    "clickhouse-driver>=0.2.10",
    "confluent-kafka>=2.3.0",
    "pyarrow>=16.0.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backfill", "producer"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
